+++ projects/loading-skeleton/src/lib/loading-skeleton.component.ts
@@ -1,19 +1,24 @@
 import { Component, OnInit, TemplateRef, Input } from "@angular/core";
 import { Observable } from "rxjs";
 import { LoadingSkeletonService } from "./loading-skeleton.service";
-import { startWith, delay } from "rxjs/operators";
-import { ITheme, ILoadingConfigTheme } from "./loading-skeleton.config";
+import { ILoadingConfigTheme } from "./loading-skeleton.config";
+import { trigger, transition, style, animate } from "@angular/animations";
 
 @Component({
   selector: "loading-skeleton",
   templateUrl: "./loading-skeleton.component.html",
-  styles: [
-    `
-      .loading-container {
-        display: flex;
-        flex-direction: column;
-      }
-    `,
+  animations: [
+    trigger("fadeIn", [
+      transition(":enter", [
+        style({ opacity: 0 }),
+        animate(
+          "300ms ease-in",
+          style({
+            opacity: 1,
+          })
+        ),
+      ]),
+    ]),
   ],
 })
 export class LoadingSkeletonComponent implements OnInit {
@@ -26,6 +31,7 @@ export class LoadingSkeletonComponent implements OnInit {
       this._theme = val;
     }
   }
+  @Input() bind: LoadingSkeletonService;
   get theme() {
     return this._theme;
   }
@@ -37,8 +43,6 @@ export class LoadingSkeletonComponent implements OnInit {
   }
   @Input() isVisible: boolean = false;
 
-  @Input() service: LoadingSkeletonService;
-
   // For debug
   // in browser: ng.getComponent($0).log()
   @Input() log = () =>
@@ -48,18 +52,27 @@ export class LoadingSkeletonComponent implements OnInit {
     });
 
   loading$: Observable<boolean>;
+  service: LoadingSkeletonService;
 
   constructor(private loadingService: LoadingSkeletonService) {}
 
   ngOnInit(): void {
-    if (this.service) {
-      this.loading$ = this.service.loading$;
-      return;
-    } /*
-    this.loading$ = this.loadingService.loading$.pipe(
-      startWith(false),
-      delay(0)
-    );*/
+    this.service = this.getService();
+    this.loading$ = this.service.loading$;
+  }
+
+  getService() {
+    return this.bind || this.loadingService;
+  }
+
+  show() {
+    this.service.show();
+    this.isVisible = true;
+  }
+
+  hide() {
+    this.service.hide();
+    this.isVisible = false;
   }
 
   get loadingContext() {