+++ projects/loading-skeleton/src/public-api.ts
@@ -2,18 +2,17 @@
  * Public API Surface of loading-skeleton
  */
 
-export * from "./lib/loading-list/loading-skeleton-list.component";
-
 export * from "./lib/loading-skeleton.service";
 export * from "./lib/loading-skeleton.component";
 export * from "./lib/loading-skeleton.module";
 export * from "./lib/loading-skeleton.config";
 
+export * from "./lib/loading-skeleton-list/loading-skeleton-list.component";
+
 export * from "./lib/loading-placeholder/loading-placeholder.component";
 
 export * from "./lib/template-components/loading-button/loading-button.component";
+export * from "./lib/template-components/loading-bullet/loading-bullet.component";
 export * from "./lib/template-components/loading-headline/loading-headline.component";
-export * from "./lib/template-components/loading-image/loading-image.component";
 export * from "./lib/template-components/loading-list/loading-list.component";
 export * from "./lib/template-components/loading-text/loading-text.component";
-export * from "./lib/template-components/loading-textarea/loading-textarea.component";