+++ src/app/app.module.ts
@@ -1,29 +1,31 @@
 import { BrowserModule } from "@angular/platform-browser";
 import { NgModule } from "@angular/core";
-
+import { FormsModule } from "@angular/forms";
+import { RouterModule } from "@angular/router";
 import { AppComponent } from "./app.component";
 import { CommonModule } from "@angular/common";
 import { NgxLoadingSkeletonModule } from "projects/loading-skeleton/src/public-api";
+import { AvailableComponent } from "./available/available.component";
+import { OptimizeComponent } from "./optimize/optimize.component";
+import { ExperimentalComponent } from "./experimental/experimental.component";
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [
+    AppComponent,
+    AvailableComponent,
+    OptimizeComponent,
+    ExperimentalComponent,
+  ],
   imports: [
     BrowserModule,
     CommonModule,
-    //NgxLoadingSkeletonModule,
-    NgxLoadingSkeletonModule.forRoot({
-      animationSpeed: "1s",
-      busyDelayMs: 300,
-      busyMinDurationMs: 3000,
-      theme: {
-        light: {
-          backgroundColor: "pink",
-        },
-        dark: {
-          backgroundColor: "red",
-        },
-      },
-    }),
+    FormsModule,
+    RouterModule.forRoot([
+      { path: "", component: AvailableComponent },
+      { path: "optimize", component: OptimizeComponent },
+      { path: "experimental", component: ExperimentalComponent },
+    ]),
+    NgxLoadingSkeletonModule,
   ],
   providers: [],
   bootstrap: [AppComponent],